Bug 803747 - private-browsing test in test-windows-common.js is throwing an error
